separate tests for Code Container
